move existing Coordinator info a frost module since we will also have fire/faire/roast
